+++ docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.1.0]
+
+**Features**
+
+- new hexadecimal methods:
+  - normal: `hex2dec`, `dec2hex`, `to_hex`, `from_hex`
+  - in-place: `hex2dec!`, `dec2hex!`, `to_hex!`, `from_hex!`
+
+**Code**
+
+- `to_b64!` and `from_b64!` code simplification
+- yard internal references from implicit to explicit
+
+**Chore**
+
+- Updated dependencies
+- Updated rubocop rules
+
 ## [1.0.0]
 
 - Initial version